+++ projeto/projeto-tcc/src/main/java/br/com/medicamento/insumo/dao/TipoControleDAO.java
@@ -17,7 +17,7 @@ public class TipoControleDAO {
 	@PersistenceContext
 	private EntityManager em;
 	
-	public Integer salvar(TipoControle tipoControle) {
+	public Long salvar(TipoControle tipoControle) {
 		em.persist(tipoControle);
 		em.flush();
 		em.refresh(tipoControle);
